Move RepositoryProvider into a different file
